+++ src/votekit/cvr_loaders/__init__.py
@@ -1,0 +1,2 @@
+from .load_csv import load_csv  # noqa
+from .load_scottish import load_scottish  # noqa
