+++ tests/cvr_loaders/test_load_csv.py
@@ -3,21 +3,12 @@ from pathlib import Path
 import pytest
 
 from votekit.ballot import Ballot
-from votekit.cvr_loaders import load_csv, load_scottish
+from votekit.cvr_loaders import load_csv
 from votekit.pref_profile import PreferenceProfile
 
 
-BASE_DIR = Path(__file__).resolve().parent
+BASE_DIR = Path(__file__).resolve().parent.parent
 CSV_DIR = BASE_DIR / "data/csv/"
-
-
-def is_equal(b1: list[Ballot], b2: list[Ballot]) -> bool:
-    if len(b1) != len(b2):
-        return False
-    for b in b1:
-        if b not in b2:
-            return False
-    return True
 
 
 def test_empty_csv():
@@ -29,8 +20,7 @@ def test_undervote():
     prof = load_csv(CSV_DIR / "undervote.csv", id_col=0)
     a = Ballot(ranking=[{"c"}, frozenset(), frozenset()], weight=1, voter_set={"a"})
     correct_prof = PreferenceProfile(ballots=[a])
-    assert correct_prof.ballots == prof.ballots
-    # assert correct_prof.ballots[0].ranking == prof.ballots[0].ranking
+    assert correct_prof == prof
 
 
 def test_only_cols():
@@ -50,14 +40,14 @@ def test_duplicates_candidates():
     don = Ballot(ranking=[{"a"}, {"c"}, {"c"}], weight=1, voter_set={"don"})
     carrie = Ballot(ranking=[{"c"}, {"c"}, {"c"}], weight=1, voter_set={"carrie"})
     correct_prof = PreferenceProfile(ballots=[abe, don, carrie])
-    assert is_equal(prof.ballots, correct_prof.ballots)
+    assert prof == correct_prof
 
 
 def test_single_row():
     prof = load_csv(CSV_DIR / "single_row.csv", id_col=0)
     a = Ballot(ranking=[{"b"}, {"c"}, {"d"}], weight=1, voter_set={"a"})
     correct_prof = PreferenceProfile(ballots=[a])
-    assert is_equal(prof.ballots, correct_prof.ballots)
+    assert prof == correct_prof
 
 
 def test_multiple_undervotes():
@@ -73,7 +63,7 @@ def test_multiple_undervotes():
         voter_set={"dave"},
     )
     correct_prof = PreferenceProfile(ballots=[abc, dave])
-    assert is_equal(correct_prof.ballots, prof.ballots)
+    assert correct_prof == prof
 
 
 def test_different_undervotes():
@@ -82,7 +72,7 @@ def test_different_undervotes():
     b = Ballot(ranking=[frozenset(), {"d"}, frozenset()], weight=1, voter_set={"b"})
     c = Ballot(ranking=[{"e"}, frozenset(), {"e"}], weight=1, voter_set={"c"})
     correct_prof = PreferenceProfile(ballots=[a, b, c])
-    assert is_equal(correct_prof.ballots, prof.ballots)
+    assert correct_prof == prof
 
 
 def test_duplicate_ballots():
@@ -90,7 +80,7 @@ def test_duplicate_ballots():
     a = Ballot(ranking=[{"b"}, {"c"}, {"c"}], weight=1, voter_set={"abe"})
     dc = Ballot(ranking=[{"c"}, {"c"}, {"c"}], weight=2, voter_set={"don", "carrie"})
     correct_prof = PreferenceProfile(ballots=[a, dc])
-    assert is_equal(correct_prof.ballots, prof.ballots)
+    assert correct_prof == prof
 
 
 def test_combo():
@@ -106,7 +96,7 @@ def test_combo():
         voter_set={"don", "ed"},
     )
     correct_prof = PreferenceProfile(ballots=[abc, de])
-    assert is_equal(correct_prof.ballots, prof.ballots)
+    assert correct_prof == prof
 
 
 def test_diff_candidates():
@@ -115,7 +105,7 @@ def test_diff_candidates():
     don = Ballot(ranking=[{"d"}, {"e"}, {"f"}], weight=1, voter_set={"don"})
     carrie = Ballot(ranking=[{"g"}, {"h"}, {"i"}], weight=1, voter_set={"carrie"})
     correct_prof = PreferenceProfile(ballots=[abe, don, carrie])
-    assert is_equal(correct_prof.ballots, prof.ballots)
+    assert correct_prof == prof
 
 
 def test_same_candidates():
@@ -124,7 +114,7 @@ def test_same_candidates():
     don = Ballot(ranking=[{"c"}, {"b"}, {"a"}], weight=1, voter_set={"don"})
     carrie = Ballot(ranking=[{"a"}, {"c"}, {"b"}], weight=1, voter_set={"carrie"})
     correct_prof = PreferenceProfile(ballots=[abe, don, carrie])
-    assert is_equal(correct_prof.ballots, prof.ballots)
+    assert correct_prof == prof
 
 
 def test_special_char():
@@ -132,7 +122,7 @@ def test_special_char():
     a1 = Ballot(ranking=[{"b@#"}, {"@#$"}, {"c"}], weight=2, voter_set={"a@#", "1@#"})
     d = Ballot(ranking=[{"!23"}, {"c"}, {"c"}], weight=1, voter_set={"d#$"})
     correct_prof = PreferenceProfile(ballots=[a1, d])
-    assert is_equal(correct_prof.ballots, prof.ballots)
+    assert correct_prof == prof
 
 
 def test_unnamed_ballot():
@@ -149,84 +139,3 @@ def test_same_name():
 #     p = CVRLoader(load_func=rank_column_csv)
 #     # p.load_cvr(DATA_DIR / "malformed.csv")
 #     # print(p)
-
-
-def test_scot_csv_parse():
-    pp, seats, cand_list, cand_to_party, ward = load_scottish(
-        CSV_DIR / "scot_wardy_mc_ward.csv"
-    )
-
-    assert seats == 1
-    assert isinstance(pp, PreferenceProfile)
-    assert set(["Paul", "George", "Ringo"]) == set(pp.candidates)
-    assert len(pp.candidates) == 3
-    assert cand_list == ["Paul", "George", "Ringo"]
-    assert cand_to_party == {
-        "Paul": "Orange (O)",
-        "George": "Yellow (Y)",
-        "Ringo": "Red (R)",
-    }
-    assert ward == "Wardy McWard Ward"
-    assert int(pp.total_ballot_wt) == 146
-    assert Ballot(ranking=tuple([frozenset({"Paul"})]), weight=126) in pp.ballots
-    assert (
-        Ballot(
-            ranking=tuple(
-                [frozenset({"Ringo"}), frozenset({"George"}), frozenset({"Paul"})]
-            ),
-            weight=1,
-        )
-        in pp.ballots
-    )
-
-
-def test_scot_csv_blank_rows():
-    pp, seats, cand_list, cand_to_party, ward = load_scottish(
-        CSV_DIR / "scot_blank_rows.csv"
-    )
-
-    assert seats == 1
-    assert isinstance(pp, PreferenceProfile)
-    assert set(["Paul", "George", "Ringo"]) == set(pp.candidates)
-    assert len(pp.candidates) == 3
-    assert cand_list == ["Paul", "George", "Ringo"]
-    assert cand_to_party == {
-        "Paul": "Orange (O)",
-        "George": "Yellow (Y)",
-        "Ringo": "Red (R)",
-    }
-    assert ward == "Wardy McWard Ward"
-    assert int(pp.total_ballot_wt) == 146
-    assert Ballot(ranking=tuple([frozenset({"Paul"})]), weight=126) in pp.ballots
-    assert (
-        Ballot(
-            ranking=tuple(
-                [frozenset({"Ringo"}), frozenset({"George"}), frozenset({"Paul"})]
-            ),
-            weight=1,
-        )
-        in pp.ballots
-    )
-
-
-def test_bad_file_path_scot_csv():
-    with pytest.raises(FileNotFoundError):
-        load_scottish("")
-
-
-def test_empty_file_scot_csv():
-    with pytest.raises(EmptyDataError):
-        load_scottish(CSV_DIR / "scot_empty.csv")
-
-
-def test_bad_metadata_scot_csv():
-    with pytest.raises(DataError):
-        load_scottish(CSV_DIR / "scot_bad_metadata.csv")
-
-
-def test_incorrect_metadata_scot_csv():
-    with pytest.raises(DataError):
-        load_scottish(CSV_DIR / "scot_candidate_overcount.csv")
-
-    with pytest.raises(DataError):
-        load_scottish(CSV_DIR / "scot_candidate_undercount.csv")
